+++ generators/graph/nodes/noise.h
@@ -46,6 +46,19 @@ void add_fast_noise_lite_state_config(ShaderGenContext &ctx, const ZN_FastNoiseL
 			fnl.get_fractal_gain(), 1.0 / fnl.get_period(), fnl.get_fractal_lacunarity());
 }
 
+void add_fast_noise_lite_gradient_state_config(ShaderGenContext &ctx, const ZN_FastNoiseLiteGradient &fnl) {
+	ctx.add_format("fnl_state state = fnlCreateState({});\n"
+				   "state.domain_warp_type = {};\n"
+				   "state.domain_warp_amp = {};\n"
+				   "state.fractal_type = {};\n"
+				   "state.octaves = {};\n"
+				   "state.gain = {};\n"
+				   "state.frequency = {};\n"
+				   "state.lacunarity = {};\n",
+			fnl.get_seed(), fnl.get_noise_type(), fnl.get_amplitude(), fnl.get_fractal_type(),
+			fnl.get_fractal_octaves(), fnl.get_fractal_gain(), 1.0 / fnl.get_period(), fnl.get_fractal_lacunarity());
+}
+
 void register_noise_nodes(Span<NodeType> types) {
 	using namespace math;
 
@@ -105,6 +118,13 @@ void register_noise_nodes(Span<NodeType> types) {
 			if (fnl.is_null()) {
 				ctx.make_error(String(ZN_TTR("Shader generation with {0} is not supported."))
 									   .format(varray(noise->get_class())));
+				return;
+			}
+			if (fnl->is_domain_warp_enabled()) {
+				// TODO Support domain warp shader generation with Godot's implementation of FastNoiseLite
+				ctx.make_error(
+						String(ZN_TTR("Shader generation from {0} with domain warp is not supported. Use {1}."))
+								.format(varray(noise->get_class(), ZN_FastNoiseLiteGradient::get_class_static())));
 				return;
 			}
 			ctx.require_lib_code("vg_fnl", zylann::fast_noise_lite::GDSHADER_SOURCE);
@@ -172,6 +192,13 @@ void register_noise_nodes(Span<NodeType> types) {
 			if (fnl.is_null()) {
 				ctx.make_error(String(ZN_TTR("Shader generation with {0} is not supported."))
 									   .format(varray(noise->get_class())));
+				return;
+			}
+			if (fnl->is_domain_warp_enabled()) {
+				// TODO Support domain warp shader generation with Godot's implementation of FastNoiseLite
+				ctx.make_error(
+						String(ZN_TTR("Shader generation from {0} with domain warp is not supported. Use {1}."))
+								.format(varray(noise->get_class(), ZN_FastNoiseLiteGradient::get_class_static())));
 				return;
 			}
 			ctx.require_lib_code("vg_fnl", zylann::fast_noise_lite::GDSHADER_SOURCE);
@@ -352,6 +379,23 @@ void register_noise_nodes(Span<NodeType> types) {
 			ctx.set_output(0, r.x);
 			ctx.set_output(1, r.y);
 		};
+
+		t.shader_gen_func = [](ShaderGenContext &ctx) {
+			Ref<ZN_FastNoiseLiteGradient> noise = ctx.get_param(0);
+			if (noise.is_null()) {
+				ctx.make_error(String(ZN_TTR("{0} instance is null"))
+									   .format(varray(ZN_FastNoiseLiteGradient::get_class_static())));
+				return;
+			}
+			ctx.require_lib_code("vg_fnl", zylann::fast_noise_lite::GDSHADER_SOURCE);
+			add_fast_noise_lite_gradient_state_config(ctx, **noise);
+			ctx.add_format("float wx = {};\n"
+						   "float wy = {};\n"
+						   "fnlDomainWarp2D(state, wx, wy);\n"
+						   "{} = wx;\n"
+						   "{} = wy;\n",
+					ctx.get_input_name(0), ctx.get_input_name(1), ctx.get_output_name(0), ctx.get_output_name(1));
+		};
 	}
 	{
 		struct Params {
@@ -412,6 +456,26 @@ void register_noise_nodes(Span<NodeType> types) {
 			ctx.set_output(0, r.x);
 			ctx.set_output(1, r.y);
 			ctx.set_output(2, r.z);
+		};
+
+		t.shader_gen_func = [](ShaderGenContext &ctx) {
+			Ref<ZN_FastNoiseLiteGradient> noise = ctx.get_param(0);
+			if (noise.is_null()) {
+				ctx.make_error(String(ZN_TTR("{0} instance is null"))
+									   .format(varray(ZN_FastNoiseLiteGradient::get_class_static())));
+				return;
+			}
+			ctx.require_lib_code("vg_fnl", zylann::fast_noise_lite::GDSHADER_SOURCE);
+			add_fast_noise_lite_gradient_state_config(ctx, **noise);
+			ctx.add_format("float wx = {};\n"
+						   "float wy = {};\n"
+						   "float wz = {};\n"
+						   "fnlDomainWarp3D(state, wx, wy, wz);\n"
+						   "{} = wx;\n"
+						   "{} = wy;\n"
+						   "{} = wz;\n",
+					ctx.get_input_name(0), ctx.get_input_name(1), ctx.get_input_name(2), ctx.get_output_name(0),
+					ctx.get_output_name(1), ctx.get_output_name(2));
 		};
 	}
 #ifdef VOXEL_ENABLE_FAST_NOISE_2